feat: error style of card component & fix useState
