+++ src/components/Card/Card.module.css
@@ -8,3 +8,7 @@
   padding-top: 22px;
   position: relative;
 }
+
+.error {
+  border: 1px solid var(--red-500);
+}
