+++ src/pages/servey/second/page.tsx
@@ -2,97 +2,26 @@ import BaseLayout from "@/pages/layouts/BaseLayout";
 import ServeyTitle from "../components/ServeyTitle";
 import Card from "@/components/Card/Card";
 import styles from "./second.module.css";
-import * as router from "@/lib/router";
-import { useEffect, useState } from "@/lib/dom";
-import FormRepository from "@/repository/FormRepository";
-
-type Select = { text: string; value: string | null; selected: boolean };
-
-const selectListInit: Select[] = [
-  {
-    text: "선택",
-    value: null,
-    selected: true,
-  },
-  {
-    text: "select option1",
-    value: "select option1",
-    selected: false,
-  },
-  {
-    text: "select option2",
-    value: "select option2",
-    selected: false,
-  },
-  {
-    text: "select option3",
-    value: "select option3",
-    selected: false,
-  },
-];
+import useSecondPageViewModel from "./hooks/useSecondPageViewModel";
 
 const SecondPage = () => {
-  const [selectList, setSelectList] = useState<Select[]>(selectListInit);
-  const [text, setText] = useState<string>("");
+  const {
+    handleSubmit,
+    isSelectedValue,
+    handleChangeSelect,
+    selectList,
+    isTextValue,
+    text,
+    handleInput,
+    goBack,
+    removeAllInputValue,
+  } = useSecondPageViewModel();
 
-  const goBack = () => {
-    router.back();
-  };
-
-  const formValidation = () => {
-    const isSelectedValue = selectList.some((select) => select.selected);
-    const isTextValue = text;
-    return isSelectedValue && isTextValue;
-  };
-
-  const handleSubmit = (e: Event) => {
-    e.preventDefault();
-    console.log("submit!!");
-    if (formValidation()) {
-      router.push("/servey/complete");
-    }
-  };
-
-  useEffect(() => {
-    const selectValue = FormRepository.getSelectBoxOption();
-    setSelectList(
-      selectList.map((select) => ({
-        ...select,
-        selected: select.value === selectValue,
-      }))
-    );
-
-    const inputValue = FormRepository.getTextAreaInput();
-    setText(inputValue);
-  }, []);
-
-  const handleChangeSelect = (e: InputEvent) => {
-    const value = (e.target as any).value;
-    setSelectList(
-      selectList.map((select) => ({
-        ...select,
-        selected: select.value === value,
-      }))
-    );
-    FormRepository.setSelectBoxOption(value);
-  };
-
-  const handleInput = (e: InputEvent) => {
-    const value = (e.target as any).value;
-    setText(value);
-    FormRepository.setTextAreaInput(value);
-  };
-  const removeAllInputValue = () => {
-    setSelectList(selectListInit);
-    setText("");
-    FormRepository.clearSelectBoxOption();
-    FormRepository.clearTextAreaInput();
-  };
   return (
     <BaseLayout>
       <form onsubmit={handleSubmit}>
         <ServeyTitle />
-        <Card>
+        <Card isError={!isSelectedValue}>
           <div className={styles.borderTop}>
             <span>section2</span>
           </div>
@@ -110,7 +39,7 @@ const SecondPage = () => {
             </select>
           </div>
         </Card>
-        <Card>
+        <Card isError={!isTextValue}>
           <div className={styles.textAreaTitle}>
             <span>textarea</span>
             <span className={styles.asterisk}>*</span>
